html/dom/elements/global-attributes/data_unicode.html

<!DOCTYPE HTML>
<html >
 <head>
  <meta charset="utf-8">
  <title> New Document </title>
  <meta name="Generator" content="EditPlus">
  <meta name="Author" content="">
  <meta name="Keywords" content="">
  <meta name="Description" content="">
<script src="testharness.js"></script>
<script src="testharnessreport.js"></script>

 </head>

 <body>
  
<div id="spaceship" data-ship-id="92432"
     data-weapons="laser 2" data-shields="50%"
     data-x="30" data-y="10" data-z="90" data-bbb="中文" >
 <!--
 <button class="fire"
         onclick="alert(this.parentNode.dataset.bbb)">
  Fire
 </button>
 -->
</div>
 <script>

 test(function() {
    var spaceship = document.getElementById("spaceship");
    assert_equals(spaceship.dataset.bbb, "中文");
	 }, "Test dataset.bbb equals to '中文'");

 

 </script>

<div id='log'></div>
 </body>
</html>
